refactor: chatting-server를 chatting으로 이름 수정
