+++ server/src/chatting/chatting.module.ts
@@ -1,5 +1,5 @@
 import { Module } from '@nestjs/common';
-import { ChattingService } from './chatting-server.service';
+import { ChattingService } from './chatting.service';
 import { StudyRoomsService } from 'src/study-room/study-room.service';
 
 @Module({
