+++ server/src/signaling-server/signaling-server.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { SignalingServerGateway } from './signaling-server.gateway';
 import { StudyRoomModule } from '../study-room/study-room.module';
-import { ChattingService } from 'src/chatting-server/chatting-server.service';
+import { ChattingService } from 'src/chatting/chatting.service';
 
 @Module({
   imports: [StudyRoomModule],
